# src/main/resources/api.yml
openapi: 3.0.3
info:
  title: Restaurant BFF OpenAPI document
  version: 0.0.1
paths:
...
  /restaurant/{restaurantId}:
    get:
      summary: Get Restaurant By id.
      description: Retrieves a restaurant by passing an id.
      operationId: getRestaurant
      parameters:
        - name: restaurantId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: "Restaurant got."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestaurantDTO"
    put:
      summary: Edit Restaurant By id.
      description: Edit restaurant by passing an id.
      operationId: editRestaurant
      parameters:
        - name: restaurantId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RestaurantDTO"
      responses:
        "200":
          description: "Restaurant edited."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestaurantDTO"
    delete:
      summary: Delete Restaurant By id.
      description: Retrieves the deleted restaurant.
      operationId: deleteRestaurant
      parameters:
        - name: restaurantId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: "Restaurant deleted."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestaurantDTO"
  /restaurant:
    post:
      summary: Create Restaurant.
      description: Create Restaurant.
      operationId: createRestaurant
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RestaurantDTO"
      responses:
        "200":
          description: "Restaurant created."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestaurantDTO"
    get:
      summary: Get All Restaurants.
      description: Get All Restaurants.
      operationId: getAllRestaurants
      responses:
        "200":
          description: "Restaurants got."
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RestaurantDTO"
  /restaurant/add/employee/{restaurantId}/{employeeId}:
    put:
      summary: Add Employee to Restaurant.
      description: Add Employee to Restaurant.
      operationId: addEmployee
      parameters:
        - name: restaurantId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: employeeId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: "Restaurant with employee added."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestaurantDTO"
  /restaurant/remove/employee/{restaurantId}/{employeeId}:
    put:
      summary: Remove Employee from Restaurant.
      description: Remove Employee from Restaurant.
      operationId: removeEmployee
      parameters:
        - name: restaurantId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: employeeId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: "Restaurant with employee removed."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestaurantDTO"
  /restaurant/{lat}/{lng}:
    get:
      summary: Persist Restaurants Near the Coordinate.
      description: Retrieves a restaurant list persisted after passing a latitude and a longitude.
      operationId: persistRestaurantsNearTheCoordinate
      parameters:
        - name: lat
          in: path
          required: true
          schema:
            type: string
        - name: lng
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: "Restaurant List Near the coordinate."
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RestaurantDTO"
  #EMPLOYEE API
...
servers:
  - url: http://localhost:8080/api
components:
  schemas:
    AddressDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        postalCode:
          type: string
        city:
          type: string
        country:
          type: string
    CoordinateDTO:
      type: object
      properties:
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
    LocationDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        address:
          $ref: "#/components/schemas/AddressDTO"
        locationCoordinate:
          $ref: "#/components/schemas/CoordinateDTO"
    MenuDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        language:
          type: string
...
    RestaurantDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        hereId:
          type: string
        name:
          type: string
        location:
          $ref: "#/components/schemas/LocationDTO"
        capacity:
          type: integer
        menu:
          $ref: "#/components/schemas/MenuDTO"
        schedule:
          $ref: "#/components/schemas/ScheduleRoutineDTO"
        employeeList:
          type: setOfEmployeeDTO
    ScheduleRoutineDTO:
      type: object
      properties:
        daysScheduleMap:
          additionalProperties:
            type: daysScheduleMap
...